AQ20/Ossirian: Add repeating event for Cyclone
this enables BigWigs to show a timer even for every fourth Cyclone.
Normally, no timer is shown after every 4th Cyclone because the tank is
not affected by the Cyclone at all due to diminishing return mechanics.
This DR-immune does not produce an event, thus the timer. Accuracy is not
negatively impacted, because Bar and Timer are refreshed on every Cyclone
sync.